agregar mensajed de confirmación; solucionar problema con paginador }

+++ src/app/maintenance/pages/maintenance-services-page/maintenance-services-page.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { FormControl } from '@angular/forms';
+import { MatDialog } from '@angular/material/dialog';
 import { PageEvent } from '@angular/material/paginator';
 import { IFilterCommon } from 'src/app/shared/filter-common';
+import { ConfirmationDialog } from '../../dialogs/confirmation-dialog/confirmation-dialog.component';
 import { ITruckMaintenanceService } from '../../interfaces/truck-maintenance-service';
 import { TruckMaintenanceServiceSerivce } from '../../services/truck-maintenance-service.service';
 
@@ -19,35 +21,42 @@ export class MaintenanceServicesPage implements OnInit {
     filter:'',
     orderBy:'',
     ascending:false,
-    page:1,
+    page:0,
     pageSize:5,
     totalRows: 0
   }
 
 
-  constructor(private truckMaintenanceService:TruckMaintenanceServiceSerivce) { }
+  constructor(private truckMaintenanceService:TruckMaintenanceServiceSerivce,
+    private dialog:MatDialog) { }
 
   ngOnInit(): void {
     this.loadData();
   }
 
   changePaginator(event:PageEvent){
+    console.log(event.pageIndex);
     this.filters.page = event.pageIndex;
     this.filters.pageSize = event.pageSize;
     this.loadData();
   }
 
+  delete(id:number){
+    const dialogRef = this.dialog.open(ConfirmationDialog, {data: {title: 'Delete Maintenance', message:'Are you sure to delete this maintenance?'}})
+    
+    dialogRef.afterClosed().subscribe(result =>{
+      if(result){
+        this.truckMaintenanceService.delete(id).subscribe(resp => {
+          this.loadData();
+        });
+      }
+    });
+  }
+  
   private loadData() {
     this.truckMaintenanceService.get(this.filters).subscribe(({body, count}) => {
       this.dataSource = body;
       this.filters.totalRows = +count;
     });
   }
-
-  delete(id:number){
-    this.truckMaintenanceService.delete(id).subscribe(resp => {
-      this.loadData();
-    });
-  }
-
 }
